Work on improving codeclimate grade.
