+++ package.json
@@ -4,7 +4,7 @@
 	"version": "2.1.3",
 	"main": "index.js",
 	"files": [
-		"index.js"
+		"*.js"
 	],
 	"scripts": {
 		"test": "xo && nyc ava -v"
